Addtask, showtasks, deltask; QoL changes
added alpha versions of addtask & showtasks and to-be-coded deltask.

Updated reload command reply to be ephemeral.

converted some functionality in addreminder and showreminders to standalone functions to be used in the task system, since the code can be reused from the reminder system, since these 2 systems are pretty similar.

Will continue to reuse code so my code is half decent, i hope